+++ src/osgview/GUIOSGView.cpp
@@ -174,7 +174,7 @@ GUIOSGView::GUIOSGView(
     myRoot = GUIOSGBuilder::buildOSGScene(myGreenLight, myYellowLight, myRedLight, myRedYellowLight, myPoleBase);
     myPlane = new osg::MatrixTransform();
     myPlane->setCullCallback(new ExcludeFromNearFarComputationCallback());
-    myPlane->addChild(GUIOSGBuilder::buildPlane((zFar - zNear)));
+    myPlane->addChild(GUIOSGBuilder::buildPlane((float)(zFar - zNear)));
     myPlane->addUpdateCallback(new PlaneMoverCallback(myViewer->getCamera()));
     myRoot->addChild(myPlane);
